pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "coach-gym-diet"
version = "0.1.0"
description = "AI dietary coach for gym goers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "anthropic>=0.40",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pillow>=10.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography]>=3.3",
    "pyjwt[crypto]>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["apps*", "workflows*", "data*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
